Create OddOccurrencesInArray.cpp
Right now the solution is rudimentary. It does not take into account the efficiency. It is of order (SqrN) which needs to be improved. Moreover the solution break every time it finds a pair which is not the correct idea. Here the solution has to wait for an even or odd number of occurrences, before that no judgement could be made about it.